// src/app/models/admin/capphephoatdongkhoangsan/cpthamdocongtrinh.model.ts
export class OutputCpThamDoCongTrinhModel {
  idcongtrinh: string;
  sohieu: string;
  chieusau: number;
  toadox: number;
  toadoy: number;
  idcapphepthamdo: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  ghichu: string;
  siteid: string;
}

export class InputCpThamDoCongTrinhModel {
  idcongtrinh: string;
  sohieu: string;
  chieusau: number;
  toadox: number;
  toadoy: number;
  idcapphepthamdo: string;
  geom: string;
  geovn: string;
  geowgs: string;
  hequychieu: string;
  ghichu: string;
  siteid: string;
}